cdc_project_sqlite/requirements.txt
python-dotenv
snowflake-connector-python
pandas